Created components dir to make stuff a little more orderly.
