+++ submit-flags/submit-flags-app/Cargo.toml
@@ -16,3 +16,7 @@ log = "0.4.6"
 wasm-logger = "0.2.0"
 sha2 = "0.9.1"
 console_error_panic_hook = "0.1.6"
+wasm-bindgen-futures = "0.4.18"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
