+++ submit-flags/submit-flags-app/src/components/level.rs
@@ -1,5 +1,4 @@
 use log;
-use std::str;
 
 use sha2::{Sha256, Digest};
