+++ submit-flags/submit-flags-app/src/components/mod.rs
@@ -1,0 +1,2 @@
+pub mod mainpage;
+pub mod level;
